+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# __1.11.930__ __2020-12-31__
+## __AWS Service Catalog__
+  - ### Features
+    - Enhanced Service Catalog DescribeProvisioningParameters API to return new parameter constraints, i.e., MinLength, MaxLength, MinValue, MaxValue, ConstraintDescription and AllowedPattern
+
 # __1.11.929__ __2020-12-30__
 ## __Amazon ElastiCache__
   - ### Features